# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit:
    post:
      operationId: deposit
      tags: [wallet]
      summary: Deposit funds to real account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: Updated balance after deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallet/withdraw:
    post:
      operationId: withdraw
      tags: [wallet]
      summary: Withdraw funds from real account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Updated balance after withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
...
components:
  schemas:
...
    WalletBalance:
      type: object
      properties:
        realBalance:
          type: number
        demoBalance:
          type: number
        currency:
          type: string
      required: [realBalance, demoBalance, currency]
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [card, bank_transfer, crypto]
      required: [amount, method]
    WithdrawInput:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, crypto]
      required: [amount, method]
...