# lib/api-zod/src/generated/types/walletBalance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading platform API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletBalance {
  realBalance: number;
  demoBalance: number;
  currency: string;
}